+++ NEWS.md
@@ -1,5 +1,7 @@
 # proxysiphon v0.0.1a1
 
+* `NcdcRecord` `description` and `original_source_url` attributes refactored to simple string.
+* Removed `contribution_date` attribute from `NcdcRecord`.
 * Grab new field from proxy text files ("Description" and "Original_Source_URL").
 * Now pull more fields from "Publication" section, including support for multiple publicatioins.
 * `Guts.pull_section()` now returns list of lists to deal with sections that appear multiple times (i.e. "Publication").